tools/custom-organization-policy-library/test/requirements.txt
pytest==8.2.1
PyYAML==6.0.1
pytest-html==4.1.1